Namespace the custom props to prevent circular JSON error
Instead of spreading the custom props (which merges the props with
router props), namespace the prop as `custom` on both the server
and the client (to maintain isomorphism).

This will prevent the following error in `./views/Layout.jsx`:
> TypeError: Converting circular structure to JSON

This is because <Layout> is rendering `props.children` and
ReactRouter passes the nested component as children.

Update <Layout> to use `this.props.custom` instead of `this.props`.