Update README with 6th lesson wrapping up react-router client setup
The sixth lesson covers refactoring the components and routes to
finish setting up react-router.

A directory named `views/` is created to house the components.
`Component.jsx` is renamed to `Layout.jsx` and `Index.jsx` and
`About.jsx` are created. These components will serve as the
view for the index route and the about route, respectively.

Use ReactRouter's Link component for routing.

The custom props are namespaced in order to prevent a circular
JSON error.

YouTube:
- Title: Client-side routing with react-router (Part 2)
- Link: https://www.youtube.com/watch?v=9T8DeVWCXE0