Update README with 7th lesson on universal rendering with Redux
The seventh lesson goes over how to set up Redux isomorphically
(on the server as well as the client).

Save dependencies `redux` and `react-redux`.

Create a reducer that returns just the default state and create
a Redux store with the initial state of the server data.

Surround the routes (on both server and client) with <Provider>
and remove the Router's `createElement` prop.

Refactor <Layout> and wrap it with ReactRedux `connect`.

Also, update webpack config to transpile both `.js` and `.jsx`
files (JSX is used in `client.js`).

YouTube:
- Title: Universal rendering with React and Redux
- Link: https://www.youtube.com/watch?v=XXjb7C7eBa8